requirements.txt
datasets>=3.0.0
gradio>=5.0.0
huggingface-hub>=0.30.0
llama-cpp-python>=0.3.0
llama-index-core>=0.12.0
networkx>=3.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
torch>=2.3.0
torch-geometric>=2.5.0
tqdm>=4.66.0

